addPercentMtRibo {cerebroApp}	R Documentation

Add percentage of mitochondrial and ribosomal transcripts.

Description

Get percentage of transcripts of gene list compared to all transcripts per cell.

Usage


addPercentMtRibo(object, organism, gene_nomenclature)

Arguments

object
Seurat object.

organism
Organism, can be either human ('hg') or mouse ('mm'). Genes need to annotated as gene symbol, e.g. MKI67 (human) / Mki67 (mouse).

gene_nomenclature
Define if genes are saved by their name ('name'), ENSEMBL ID ('ensembl') or GENCODE ID ('gencode_v27', 'gencode_vM16').

Details

Add percentage of mitochondrial and ribosomal transcripts.

Value

Seurat object with two new meta data columns containing the percentage of mitochondrial and ribosomal gene expression for each cell.

Examples


seurat <- addPercentMtRibo(
  object = seurat,
  organism = 'hg',
  gene_nomenclature = 'name'
)

[Package cerebroApp version 1.1.0 Index]