calculatePercentGenes {cerebroApp}	R Documentation

Calculate percentage of transcripts of gene list.

Description

Get percentage of transcripts of gene list compared to all transcripts per cell.

Usage


calculatePercentGenes(object, genes)

Arguments

object
Seurat object.

genes
List(s) of genes.

Details

Calculate percentage of transcripts of gene list.

Value

List of lists containing the percentages of expression for each provided gene list.

Examples


calculatePercentGenes(
  object = seurat,
  genes = list('example' = c('Fth1','Atf1'))
)

[Package cerebroApp version 1.1.0 Index]