exportFromSeurat {cerebroApp}	R Documentation

Export Seurat object to Cerebro.

Description

This function allows to export a Seurat object to visualize in Cerebro.

Usage


exportFromSeurat(
  object,
  assay = "RNA",
  file,
  experiment_name,
  organism,
  column_sample = "sample",
  column_cluster = "cluster",
  column_nUMI = "nUMI",
  column_nGene = "nGene",
  column_cell_cycle_seurat = NULL,
  column_cell_cycle_cyclone = NULL,
  add_all_meta_data = TRUE
)

Arguments

object
Seurat object.

assay
Assay to pull counts from; defaults to 'RNA'. Only relevant in Seurat v3.0 or higher since the concept of assays wasn't implemented before.

file
Where to save the output.

experiment_name
Experiment name.

organism
Organism, e.g. hg (human), mm (mouse), etc.

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

column_nUMI
Column in object@meta.data that contains information about number of transcripts per cell; defaults to 'nUMI'.

column_nGene
Column in object@meta.data that contains information about number of expressed genes per cell; defaults to 'nGene'.

column_cell_cycle_seurat
Optional column in object@meta.data that contains information about cell cycle phase based on Regev method (default of Seurat); defaults to NULL.

column_cell_cycle_cyclone
Optional column in object@meta.data that contains information about cell cycle phase based on Cyclone method; defaults to NULL.

add_all_meta_data
If set to TRUE, all further meta data columns will be extracted as well.

Details

Export Seurat object to Cerebro.

Value

Nothing.

Examples


pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
exportFromSeurat(
  object = pbmc,
  file = 'pbmc.crb',
  experiment_name = 'PBMC',
  organism = 'hg',
  column_sample = 'sample',
  column_cluster = 'seurat_clusters',
  column_nUMI = 'nCount_RNA',
  column_nGene = 'nFeature_RNA'
)

[Package cerebroApp version 1.2.2 Index]