getMarkerGenes {cerebroApp}	R Documentation

Get marker genes for every sample and cluster in Seurat object.

Description

This function gets marker genes for every sample and cluster of the Seurat object.

Usage


getMarkerGenes(object, organism = NULL, column_sample = "sample",
  column_cluster = "cluster", only_pos = TRUE, min_pct = 0.7,
  thresh_logFC = 0.25, thresh_p_val = 0.01, test = "wilcox",
  verbose = TRUE, ...)

Arguments

object
Seurat object.

organism
Organism information for pulling info about presence of marker genes of cell surface; can be omitted if already saved in Seurat object; defaults to NULL.

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

only_pos
Identify only over-expressed genes; defaults to TRUE.

min_pct
Only keep genes that are expressed in at least n% of current group of cells, defaults to 0.70 (70%).

thresh_logFC
Only keep genes that show an average logFC of at least n; defaults to 0.25.

thresh_p_val
Threshold for p-value, defaults to 0.01.

test
Statistical test used, defaults to 'wilcox' (Wilcoxon test).

verbose
Print progress bar; defaults to TRUE.

...
Further parameters can be passed to control Seurat::FindAllMakers().

Details

Get marker genes for every sample and cluster in Seurat object.

Value

Seurat object with marker gene results for samples and clusters stored in object@misc$marker_genes.

Examples


seurat <- getMarkerGenes(
  object = seurat
  organism = 'hg',
  column_sample = 'sample',
  column_cluster = 'cluster',
  only_pos = TRUE,
  min_pct = 0.7,
  thresh_logFC = 0.25,
  thresh_p_val = 0.01,
  test = 'wilcox',
  verbose = TRUE
)

[Package cerebroApp version 1.1.0 Index]