getMostExpressedGenes {cerebroApp}	R Documentation

Get most expressed genes for every sample and cluster in Seurat object.

Description

This function calculates the most expressed genes for every sample and cluster of the Seurat object.

Usage


getMostExpressedGenes(object, column_sample = "sample",
  column_cluster = "cluster")

Arguments

object
Seurat object.

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

Details

Get most expressed genes for every sample and cluster in Seurat object.

Value

Seurat object with most expressed genes stored for every sample and cluster stored in object@misc$most_expressed_genes.

Examples


seurat <- getMostExpressedGenes(
  object = seurat,
  column_sample = 'sample',
  column_cluster = 'cluster'
)

[Package cerebroApp version 1.1.0 Index]