read_GMT_file {cerebroApp}	R Documentation

Read GMT file.

Description

This functions reads a (tab-delimited) GMT file which contains the gene set name in the first column, the gene set description in the second column, and the gene names in the following columns.

Usage


read_GMT_file(file)

Arguments

file
Path to GMT file.

Details

Read GMT file.

Value

Returns an object in the same format as from the GSA.read.gmt function (GSA package) with gene sets, gene set names, and gene set descriptions stored in lists.

Examples


my_gmt <- read_GMT_file(
  file = 'path/to/my/gene_sets.gmt'
)

[Package cerebroApp version 1.1.0 Index]