calculatePercentGenes {cerebroApp}	R Documentation

Calculate percentage of transcripts of gene list.

Description

Get percentage of transcripts of gene list compared to all transcripts per cell.

Usage


calculatePercentGenes(object, assay = "RNA", genes)

Arguments

object
Seurat object.

assay
Assay to pull counts from; defaults to 'RNA'. Only relevant in Seurat v3.0 or higher since the concept of assays wasn't implemented before.

genes
List(s) of genes.

Details

Calculate percentage of transcripts of gene list.

Value

List of lists containing the percentages of expression for each provided gene list.

Examples


pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
calculatePercentGenes(
  object = pbmc,
  genes = list('example' = c('FCN1','CD3D'))
)

[Package cerebroApp version 1.2.2 Index]