Get marker genes for every sample and cluster in Seurat object.

This function gets marker genes for every sample and cluster of the Seurat object.

getMarkerGenes(
  object,
  assay = "RNA",
  organism = NULL,
  column_sample = "sample",
  column_cluster = "cluster",
  only_pos = TRUE,
  min_pct = 0.7,
  thresh_logFC = 0.25,
  thresh_p_val = 0.01,
  test = "wilcox",
  verbose = TRUE,
  ...
)

Arguments

object
Seurat object.

assay
Assay to pull counts from; defaults to 'RNA'. Only relevant in Seurat v3.0 or higher since the concept of assays wasn't implemented before.

organism
Organism information for pulling info about presence of marker genes of cell surface; can be omitted if already saved in Seurat object; defaults to NULL.

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

only_pos
Identify only over-expressed genes; defaults to TRUE.

min_pct
Only keep genes that are expressed in at least n% of current group of cells, defaults to 0.70 (70%).

thresh_logFC
Only keep genes that show an average logFC of at least n; defaults to 0.25.

thresh_p_val
Threshold for p-value, defaults to 0.01.

test
Statistical test used, defaults to 'wilcox' (Wilcoxon test).

verbose
Print progress bar; defaults to TRUE.

...
Further parameters can be passed to control Seurat::FindAllMakers().

Value

Seurat object with marker gene results for samples and clusters stored in object@misc$marker_genes.

Details

Get marker genes for every sample and cluster in Seurat object.

Examples

pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
pbmc <- getMarkerGenes(
  object = pbmc,
  organism = 'hg',
  column_sample = 'sample',
  column_cluster = 'seurat_clusters',
  only_pos = TRUE,
  min_pct = 0.7,
  thresh_logFC = 0.25,
  thresh_p_val = 0.01,
  test = 'wilcox',
  verbose = TRUE
)

#> [13:28:05] Sample column provided but only 1 sample found.

#> [13:28:05] Get marker genes for 3 clusters...

#> Calculating cluster 0

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Warning: cannot compute exact p-value with ties

#> Calculating cluster 1

#> Calculating cluster 2