getMostExpressedGenes {cerebroApp}	R Documentation

Get most expressed genes for every sample and cluster in Seurat object.

Description

This function calculates the most expressed genes for every sample and cluster of the Seurat object.

Usage


getMostExpressedGenes(
  object,
  assay = "RNA",
  column_sample = "sample",
  column_cluster = "cluster"
)

Arguments

object
Seurat object.

assay
Assay to pull counts from; defaults to 'RNA'. Only relevant in Seurat v3.0 or higher since the concept of assays wasn't implemented before.

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

Details

Get most expressed genes for every sample and cluster in Seurat object.

Value

Seurat object with most expressed genes stored for every sample and cluster stored in object@misc$most_expressed_genes.

Examples


pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
pbmc <- getMostExpressedGenes(
  object = pbmc,
  column_sample = 'sample',
  column_cluster = 'seurat_clusters'
)

[Package cerebroApp version 1.2.2 Index]