performGeneSetEnrichmentAnalysis {cerebroApp}	R Documentation

Perform gene set enrichment analysis with GSVA.

Description

This function calculates enrichment scores, p- and q-value statistics for provided gene sets for samples and clusters of cells in given Seurat object using gene set variation analysis (GSVA). Calculation of p- and q-values for gene sets is performed as done in "Evaluation of methods to assign cell type labels to cell clusters from single-cell RNA-sequencing data", Diaz-Mejia et al., F1000Research (2019).

Usage


performGeneSetEnrichmentAnalysis(
  object,
  assay = "RNA",
  GMT_file,
  column_sample = "sample",
  column_cluster = "cluster",
  thresh_p_val = 0.05,
  thresh_q_val = 0.1,
  ...
)

Arguments

object
Seurat object.

assay
Assay to pull counts from; defaults to 'RNA'. Only relevant in Seurat v3.0 or higher since the concept of assays wasn't implemented before.

GMT_file
Path to GMT file containing the gene sets to be tested. The Broad Institute provides many gene sets which can be downloaded: http://software.broadinstitute.org/gsea/msigdb/index.jsp

column_sample
Column in object@meta.data that contains information about sample; defaults to 'sample'.

column_cluster
Column in object@meta.data that contains information about cluster; defaults to 'cluster'.

thresh_p_val
Threshold for p-value, defaults to 0.05.

thresh_q_val
Threshold for q-value, defaults to 0.1.

...
Further parameters can be passed to control GSVA::gsva().

Details

Perform gene set enrichment analysis with GSVA.

Value

Seurat object with GSVA results for samples and clusters stored in object@misc$enriched_pathways$GSVA.

Examples


pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
example_gene_set <- system.file("extdata/example_gene_set.gmt",
  package = "cerebroApp")
pbmc <- performGeneSetEnrichmentAnalysis(
  object = pbmc,
  GMT_file = example_gene_set,
  column_sample = 'sample',
  column_cluster = 'seurat_clusters',
  thresh_p_val = 0.05,
  thresh_q_val = 0.1,
  parallel.sz = 1,
  verbose = FALSE
)

[Package cerebroApp version 1.2.2 Index]